% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{all_na}
\alias{all_na}
\title{function to test if an entire column is NA}
\usage{
all_na(x)
}
\arguments{
\item{x}{any column in a dataframe}
}
\value{
Returns boolean TRUE if column is all NA, FALSE if not
}
\description{
This function tests for completely NA columns
}
\examples{
test <- all_na(c(NA, "a", NA))

}
