% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{anomalies}
\alias{anomalies}
\title{Fixes the anomlies}
\usage{
anomalies(df)
}
\arguments{
\item{df}{structured dataframe}
}
\value{
df with two columns of cleaned_energy and anom_flag
}
\description{
This function gets the data and finds the anomlies in weekends and
weekdays and gives a dataframe with anomalies and anomaly columns
}
\author{
Arash Khalilnejad
}
