% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{anomaly_detector}
\alias{anomaly_detector}
\title{detects rhw anomalies and returns a dataframw with cleaned and anom_flag column}
\usage{
anomaly_detector(df, batch_days = 90)
}
\arguments{
\item{df}{the strucutred data}

\item{batch_days}{the batch of data that the anomaly detection is applied. Since time series decomposition is used,
one seasonality will be applied for whole data which is inefficeint, if NA, will pass whole}
}
\value{
data with anomalies
}
\description{
detects rhw anomalies and returns a dataframw with cleaned and anom_flag column
}
\author{
Arash Khalilnejad
}
