% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{data_structure}
\alias{data_structure}
\title{Reads jci files gotten in budget period 2}
\usage{
data_structure(df, col = "elec_cons", timestamp_col = "timestamp")
}
\arguments{
\item{df}{dataframe containing at least the timestamp column and the variable to be plotted with the heatmap}

\item{col}{the character name of the column to be ploted}

\item{timestamp_col}{the character name of the timestamp column
which i is the number of file in the list}
}
\value{
a dataframe with fixed timestamps and useful cooumns
}
\description{
Reads the jci file and modifies the timestamp intevals and
based on location modifies the timezone using googleapi and
then generates the useful columns
}
\author{
Arash
}
