% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{day_time_start_end}
\alias{day_time_start_end}
\title{finds median start and end time of PV operation}
\usage{
day_time_start_end(df)
}
\arguments{
\item{df}{with num_time Column}
}
\value{
dataframe with start and end time
}
\description{
finds median start and end time of PV operation
}
\author{
Arash Khalilnejad
}
