% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{df_With_on_time}
\alias{df_With_on_time}
\title{data with PV on time flag.}
\usage{
df_With_on_time(df)
}
\arguments{
\item{df}{df with num_time}
}
\value{
input data with one more column of on_time
}
\description{
returns dataframe of PV with approximate operating period, baed on median of start and end time.
}
\author{
Arash Khalilnejad
}
