% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{grade_pv}
\alias{grade_pv}
\title{returns quality information of time series data of PV}
\usage{
grade_pv(
  df,
  col = "poay",
  id = "pv_id",
  timestamp_col = "tmst",
  timestamp_format = "\%Y-\%m-\%d \%H:\%M:\%S",
  batch_days = 90
)
}
\arguments{
\item{df}{the PV time series data. It can be the direct output of read.csv(file_name, stringsAsFactors = F)}

\item{col}{column of the grading, default 'poay'}

\item{id}{The name of the pv data}

\item{timestamp_col}{the character name of the timestamp column}

\item{timestamp_format}{the POSIXct format of the timestamp if conversion is needed}

\item{batch_days}{the batch of data that the anomaly detection is applied. Since time series decomposition is used,
one seasonality will be applied for whole data which is inefficeint, if NA, will pass whole}
}
\description{
returns quality information of time series data of PV
}
\author{
Arash Khalilnejad
}
