% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{ip_num_time}
\alias{ip_num_time}
\title{Numerical time interim predictor.}
\usage{
ip_num_time(data, ts_col = "timestamp")
}
\arguments{
\item{data}{A dataframe with a timestamp column.}

\item{ts_col}{The timestamp column name in \code{data}. Default value is
'timestamp'.}
}
\value{
\code{data} with a num_time column added.
}
\description{
Convert the hour and minute component of each timestamp to a numerical
representation.
}
\author{
Arash Khalilnejad
}
