% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{lin_inter_hrly_to_fifteen}
\alias{lin_inter_hrly_to_fifteen}
\title{Linearly interpolate hourly data to 15 min data.}
\usage{
lin_inter_hrly_to_fifteen(data, data_ts)
}
\arguments{
\item{data}{A data frame with hourly data.}

\item{data_ts}{The column name for the \code{data}
timestamp.}
}
\value{
The resulting fifteen minute data frame.
}
\description{
Many weather data sets are hourly and we need values for
every 15 minutes.
}
\details{
Any value that can not be linearly interpolated such as a
string will remain the same.
}
\author{
Arash Khalilnejad
}
