% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{lin_inter_missing_energy}
\alias{lin_inter_missing_energy}
\title{Linearly interpolate missing energy values.}
\usage{
lin_inter_missing_energy(data, threshold = 4)
}
\arguments{
\item{data}{A data frame with an 'elec_cons' column.}

\item{threshold}{The maximum number of consective values
that may be filled with interpolated values. By default
four.}
}
\value{
The data frame with 'missing values' filled in.
}
\description{
If there exist lest than four missing values, represented by
NA values, fill with linearly interpolated values.
}
\examples{
\dontrun{
lin_inter_missing_energy(data)
}
}
