% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{nc}
\alias{nc}
\title{function to convert to character then numeric}
\usage{
nc(x)
}
\arguments{
\item{x}{any factor to convert to numeric}
}
\value{
Returns supplied parameter as numeric
}
\description{
The function is a shorthand for converting factors to numeric
}
\examples{
num <- nc(test_df$power)

}
