% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{num_test}
\alias{num_test}
\title{function to test is the values in a column should be numeric}
\usage{
num_test(col)
}
\arguments{
\item{col}{any column in a dataframe}
}
\value{
Returns boolean TRUE if column should be numeric, FALSE if not
}
\description{
This function tests a column to see if it should be numeric
}
\examples{
test <- num_test(test_df$power)

}
