% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{parallel_cluster_export}
\alias{parallel_cluster_export}
\title{Export variables to a cluster.}
\usage{
parallel_cluster_export(cluster, varlist, envir = .GlobalEnv)
}
\arguments{
\item{cluster}{Cluster}

\item{varlist}{Character vector of names of objects to export.}

\item{envir}{Environment from which t export variables}
}
\description{
Ghost cluster export call to make sure
testCoverage's trace function and environment
are available.
}
