% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{plr_convert_columns}
\alias{plr_convert_columns}
\title{Fix Column Typings}
\usage{
plr_convert_columns(df)
}
\arguments{
\item{df}{A dataframe containing pv data.}
}
\value{
Returns original dataframe with columns corrected to proper classes
}
\description{
Converts appropriate columns to numeric without specifying
 the name of the column. All columns from hbase are read as factors.
 Columns are tested to see if they should be numeric by forcing conversion
 to numeric. Columns that subsequently contain NA's are not numeric; if not,
 they are set to numeric.
}
\examples{
df <- PVplr::plr_convert_columns(test_df)

}
