% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_df.R
\docType{data}
\name{test_df}
\alias{test_df}
\title{DOE RTC Sample PV System Data}
\format{
A .csv file that can be read as a dataframe. 16265 rows and 22 variables.
}
\usage{
test_df
}
\description{
A dataset containing a small, randomly taken sample of PV data
from SDLE's data collection. It is included for the purposes of unit 
tests and vignettes, serving as an example of how the package's functions work.
}
\keyword{datasets}
