% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{time_frequency}
\alias{time_frequency}
\title{Determines the minutes between data points in a time-series}
\usage{
time_frequency(data)
}
\arguments{
\item{data}{A time-series dataframe containing a column named 'timestamp'.}
}
\value{
a numeric value of the minutes between data points
}
\description{
Determines the minutes between data points in a time-series
}
\author{
Arash Khalilnejad
}
