% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_quality.R
\name{ts_inflate}
\alias{ts_inflate}
\title{Inflate a time series data set.}
\usage{
ts_inflate(data, ts_col, col_to_imp, dt)
}
\arguments{
\item{data}{A data frame containing columns \code{ts_col} and
\code{col_to_imp}.}

\item{ts_col}{The name of the timestamp column.}

\item{col_to_imp}{The name of the column to impute.}

\item{dt}{The expected time between consecutive timestamps, in minutes.}
}
\description{
Shifts known values to the nearest equidistant timestamp and fills in any
missing timestamps with NA values. An additional binary column named
\code{<column to impute>_imp} is added where 1 represents an unknown value
and zero represents a known value.
}
